pyproject.toml
[tool.poetry]
name = "bankapp"
version = "0.1.0"
description = ""
authors = ["daveraskin <draskin1991@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "4.2.2"
asgiref = "3.7.2"
autopep8 = "2.0.2"
django-cors-headers = "4.1.0"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.2.2"
nanoid = "2.0.0"
psycopg2-binary = "2.9.9"
pycodestyle = "2.10.0"
pyjwt = "2.7.0"
python-dotenv = "1.0.0"
pytz = "2023.3"
sqlparse = "0.4.4"
tomli = "2.0.1"
typing-extensions = "4.6.3"
dj-database-url = "^2.1.0"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
